Fix bug in example (incorrect regular expression)
I updated the regular expression so it would find all tmp folders. The previous regular expression was throwing an error because `\` is the escape symbol in R strings. To include it in a regular expression, you have to put it in twice `\\`, the first one is the escape symbol and then then second one actually gets included in the string.

I also removed the don't run tag around the expression, as requested by in the CRAN submission feedback.