Remove unused function source_files()
This function was unused by any other function in the package and also wrote to the user's global environment, which is not allowed by CRAN. It's not a part of the package's core functionality and should be removed.

Also updated examples for source_funcs and source_lines to use a dummy environment.